backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langgraph>=0.0.26
python-dotenv>=1.0.0
pydantic>=2.0.0
motor>=3.0.0
